Split Unknown14 into two variables added GetFieldNameFromOffset type extension added ContainsComplextStructures attribute to some Chunk structures added non buffer ctor to SramFileSoE Bugfix SramSizes: Unknown14 was wrong

+++ Models/Enums/Unknown/Unknown14.cs
@@ -5,7 +5,7 @@ using System;
 namespace SRAM.SoE.Models.Enums.Unknown
 {
 	[Flags]
-	public enum Unknown14 : uint
+	public enum Unknown14 : UInt16
 	{
 		AnnihilationAmuletMerchantScrewed = 0x40 // Merchant for annihilation amulett blocked "you're too screwed"
 	}
